--- starsessions/__init__.py
@@ -1,21 +1,0 @@
-from .middleware import SessionMiddleware
-from .session import (
-    CookieBackend,
-    ImproperlyConfigured,
-    InMemoryBackend,
-    Session,
-    SessionBackend,
-    SessionError,
-    SessionNotLoaded,
-)
-
-__all__ = [
-    "SessionMiddleware",
-    "Session",
-    "SessionBackend",
-    "CookieBackend",
-    "InMemoryBackend",
-    "SessionNotLoaded",
-    "ImproperlyConfigured",
-    "SessionError",
-]
